update the utility functions for the lambda ingest function
