+++ requirements.txt
@@ -8,7 +8,8 @@ flake8==7.1.2
 pip-audit==2.8.0
 docformatter==1.7.5
 pytest-cov==6.0.0
-psycopg2==2.9.10
 dotenv==0.9.9
+psycopg2-binary==2.9.10
+psycopg2==2.9.10
